refactor(domain): decompose git-pr-service into 5 focused services with facade
Extract 1000+ line GitPrService god class into focused sub-services:
- DiffAnalyzerService: diff summary and unified diff parsing
- BranchDiscoveryService: remote/branch discovery, sync status
- CiStatusService: CI status, watch, failure logs, merge verification
- PrCreationService: PR creation, GitHub repo, PR status listing
- MergeStrategyService: merge, rebase, sync, conflict resolution, stash

GitPrService remains as a thin facade delegating to sub-services,
preserving backward compatibility for all existing callers including
tests that use `new GitPrService(execFn)`.
